API refactor
