+++ src/nevermined/nfts/SubscriptionNFT.ts
@@ -1,0 +1,6 @@
+import { Nft721Api } from '../Nft721Api'
+
+export default class SubscriptionNFT extends Nft721Api {
+
+
+}
